Bug fix and refactored some other uses of "long"

+++ include/vlog/seminaiver.h
@@ -143,7 +143,7 @@ class SemiNaiver {
 
         virtual FCIterator getTableFromEDBLayer(const Literal & literal);
 
-        virtual long getNLastDerivationsFromList();
+        virtual size_t getNLastDerivationsFromList();
 
         virtual void saveDerivationIntoDerivationList(FCTable *endTable);
